+++ src/monaco/__init__.py
@@ -2,6 +2,9 @@
 from importlib import metadata
 __version__ = metadata.version(__name__)
 
+import logging
+logging.getLogger(__name__).addHandler(logging.NullHandler())
+
 from monaco.mc_case import *
 from monaco.mc_sim import *
 from monaco.mc_var import *