+++ CHANGELOG.md
@@ -12,7 +12,6 @@
 - Example: Evidence-Based Scheduling
 - 2D/3D Line statistics
 - Correlated inputs
-- Switch from prints to logging
 - Ability to plot derived data in addition to vars
 - Star Sampling
 - Variogram Sensitivity Analysis
@@ -28,8 +27,10 @@
 * `SimulationFunctions` dataclass to hold the preprocess, run, and postprocess functions with stronger checks and typing. Sim can still take in the original dict of those functions with specific keys, to be backwards compatible ([GH-16](https://github.com/scottshambaugh/monaco/issues/16))
 * "Ecosystem" section of the changelog for libraries that extend monaco ([GH-17](https://github.com/scottshambaugh/monaco/issues/17))
 ### Changed    
+* The `logging` library is now used instead of print statements
 * `__repr__` updates for Sim, Case, and VarStat
 ### Removed    
+* `vprint` and `vwarn`
 
 ## [0.18.0] - 2025-07-22
 ### Added